# perf(detection): use strings.Builder